Floodfilling: support for retrieving results
Added support for retrieving the results of the flood filling and loading them
into the volume table. Note currently all results are saved, since this
may generate huge amounts of data we might want to have temporary
storage for flood filling results that would then be made permanent if
desired.